WIP: multi-OS container support
This is the first bits of a multi-OS support implementation. I don't see
any clear way to specify a base image unless than creating sub
directories. So yes, this duplicates some part of the code but given
that we use automated builds I don't see how we can use Docker's build
system to search for different 'base'.
My idea was to have something like:

FROM %%CEPH_NAMESPACE%%/%%CEPH_PREFIX%%base:%%CEPH_TAG%%

It looks like we can achieve this only by using an external build system
that will push our images to the hub. However that means having an infra
somewhere, maintain it and all the pain that comes with it. I don't
believe @Ulexus or I (@leseb) are willing to invest in such system.

This requires heavy review before merging as this will break our
building system. We need to modify the hub to reflect this change.

Signed-off-by: Sébastien Han <seb@redhat.com>