Necessary changes for k8s to work
We can't use `--net=host` with k8s. Thus if we manually set `MON_IP` as
an environement variable we will get duplicate content when using a
replication controller.
Introducing a new variable to auto detect the IP address of the
container.
This will generally work for every use case that don't want to use
`--net=host`.

Signed-off-by: Sébastien Han <seb@redhat.com>