Remove OSD variable
This variable is not used anywhere and is probably confusing for users.
We could potentially set a default for it (either `osd_ceph_disk` or
`osd_directory`). However given that both invocations are really
different we should not attempt to workaround. Thus the user has to set
the proper options.

Signed-off-by: Sébastien Han <seb@redhat.com>