entrypoint.sh: mds auth error w/ admin keyring
mds won't start for me with the following error:

2015-07-16 16:20:14.168684 7f95b6028700  0 librados: client.bootstrap-mds authentication error (1) Operation not permitted
Error connecting to cluster: PermissionError

when running 
   ceph --cluster ${CLUSTER} --name client.bootstrap-mds $KEYRING_OPT health
with $KEYRING_OPT set to
   --keyring /etc/ceph/${CLUSTER}.client.admin.keyring

It appears that the admin keyring (/var/lib/ceph/bootstrap-mds/ceph.keyring) is not accepted, however, it does work with the bootstrap-mds keyring in /var/lib/ceph/bootstrap-mds/ceph.keyring 

Is this the right approach? I've got 7 machines running each a container with mon and osd so far.