Fix to run ceph/daemon osd on Centos.
From entrypoint.sh, 'ceph-disk activate' call is made to 
activate the disk and later start the ceph-osd process
as a daemon. Currently as systemd is not initalised
as part of base Centos image, this seizes to run and errors out.

Also, another call is made from entrypoint.sh after the 
initialization to kick start ceph osd process in the 
foreground for the ceph-osd docker container to continuously run.
As a result, there are two calls made to run ceph-osd and
this will result in the container erroring out and exiting.
This patch aims to fix the same.

Signed-off-by: Deepthi Dharwar <ddharwar@redhat.com>
Signed-off-by: Sébastien Han <seb@redhat.com>