daemon: multiple osd in one container
Feature proposed by @oll-78.
It brings the ability to run multiple OSDs within a single container.
I thought we were already able to do so but didn't find any mention in
the code. The only thing I recall is a patch from @dmitryint in #24 but
since I never used it I'll let either @dmitryint or @Ulexus comment on
this. It seems that this patch uses ceph-disk instead.

Signed-off-by: Oleg Lazo <oll@niif.spb.su>
Signed-off-by: Sébastien Han <seb@redhat.com>