workflow: fix cp invocation
We need to call cp with wilcard otherwise we will end up having both
base and daemon directories content in ./
Also removing the symlinks are copying the original files instead.
Prior to this change the hub was failing because symlinks were broken so
we need to use the original files in base and daemon directories.

Signed-off-by: Sébastien Han <seb@redhat.com>