ceph-docker-workflow: Don't forget symlinked builds
Most of the actual files are in fact symlinks.
The current code was detecting changes on the original file but not on
the linked ones leading to a situation where some distro were forgoten.

Like ubuntu 16 linking ubuntu 14, a change in u14 didn't triggered a
rebuild for u16.

This patch does two things :
- creating a map between the links and the original files
- making a lookup between those impacted file and the missing distro

As a result, this patch flags the distros that have linked file as
impacted once the original is touched.