CI: build daemon image on travis and test it
We rely on Travis's infrastructure to build the 'daemon' image and run
all the ceph daemons. So we run a monitor, an osd (with directory
scenario, a mds and one rgw).
We have a validation script to make they all started properly. During
the validation sequence, we wait for 20 seconds to let the cluster
stabilize and populate the pgs.

Signed-off-by: Sébastien Han <seb@redhat.com>