daemon: new directory mode
the current directory mode aims to run multiple osd processes in a
single container. It uses runit, which is hard to find and not packaged
for some distros. This commit introduces a simple directory mode that
will only run one osd process per container. Prior to do that it will
check if a lock is already held by another osd process. If so it won't
start the osd and will just try to find another one to run. If nothing
is found or all the osds available are already running with exit the
container explaining that all the osds are already running.

Signed-off-by: Sébastien Han <seb@redhat.com>