Disable osd check max obj name len in jewel:demo
Since jewel, certain limitations apply to OSD object name
lengths: specifically if ext4 is in use for
block devices or a directory based OSD is configured, OSD's
must be configured to limit object name length:

  osd max object name len = 256
  osd max object namespace len = 64

Or disable the check:

  osd check max object name len on startup = false

When using ceph/demo, most of users are using an ext4 fs. So we should
open this setting by default.