daemon: add some verbose to ceph-disk scenario
See #326, if you run your container using `OSD_FORCE_ZAP=1` along with
the ceph_disk scenario, if you restart the container then the device
will get formatted.

Since the container keeps its properties and
`OSD_FORCE_ZAP=1` was enabled. This results in the device to be formatted.
We detect that the device is an OSD but we zap it.

If you don't specify any osd_type, the entrypoint will only activate the
device.

Signed-off-by: Sébastien Han <seb@redhat.com>