workflow: fix change selection
From now on, we grep to pin the current branch we are running the
comparaison on  so we only get the changes that are part of our branch, if
don't do this the following will occor:
infernalis will see that jewel changed and will request jewel to
change however jewel is already up to date. Then we fall into an
infite loop.

Signed-off-by: Sébastien Han <seb@redhat.com>