osd: reduce log verbosity
As described in #436, using `-d` to run the ceph-osd process which makes
docker logs growing really fast.
Using `-d` is not ideal nor really useful.

Run `ceph-osd` process with `-f` (which will log to files) instead of
running with `-d -f` which will only log on stdout and not log file.

`-f` seems to log errors only, which should be enough.

Closes: #436
Signed-off-by: Sébastien Han <seb@redhat.com>