Timeout loop to detect IP addresses of ceph-mons
Race condition at times prevents the proper detection of ceph-mon pod IP
addresses resulting in the creation of an invalid monmap. This change
implements a 10 second retry loop to attempt to grab the ceph-mon pod IP
addresses before giving up and exiting. If we exit, we allow kubernetes to
restart the pod.

Signed-off-by: Ivan Font <ifont@redhat.com>