Update CephFS and RBD tests to use nodeSelector
The CephFS and RBD tests have been updated to require nodes labeled as
a ceph-client node-type in order to be eligible to run the test pods.
This helps target client nodes which are able to have the necessary
Ceph and RBD utilities installed to run the tests. This also updates the
documentation to help clarify this.