config.kv.sh: Removing duplicated code
The removed code was in fact what is done in get_config().
So removing that lines by the get_config call makes the code easier to
read and reusing existing fucntions is always a good win.

This change will just make a second reissue of the "get montSetupComplete"
which is not a great loss while comparing with the benefits of removing duplicated code.