daemon: remove kviator
Our last survey, in #472 gave us a better understanding on how
ceph-docker is being used. If we look at the results here:
https://www.poll-maker.com/results968938xb3e2202E-41#tab-2 it seems that
consul kv store is not really popular. Over the 58 people that answered,
21 people use a kv store and only 4 people use consul. Within these 4
people, it's not clear if they use kviator to push their keys or the
consul CLI.

At the very beginning, we were providing support for etcdctl only. Now
the project is getting more attention and companies like Red Hat and
potentially Suse (due to their recent image addition) are looking at
productizing this.
For that, we would like to have as low packages dependencies as
possible. Kviator development is not really moving, maybe for a simple
reason: it does the job and there is nothing to add.

This commit re-introduces etcdctl with a pluggable interface that will
allow us to use more kv stores if we need to.
No changes were introduced in ceph-ansible, the current deploy with kv
functionality works as it used to.

As per discussion here: https://github.com/ceph/ceph-docker/issues/497

Signed-off-by: Sébastien Han <seb@redhat.com>