common_functions: creating all mandatory directories only once
The actual code is full of "mkdir /var/lib/ceph/...; chown ceph. /var/lib/ceph/..."

The idea of this patch is to relocate into a single place the creation
of all mandatory directories in a single place an call it once.

Then, it will be up to sub-programs managing theirs specific directories
but never the mandatory ones.

create_bootstrap_directories() is renamed into create_mandatory_directories() and called once at the early stage.