entrypoint: Adding DEBUG environment variable
If a container fails to start, it's pretty complicated to understand what went wrong and why.
As per the container thing, it's even more difficult as if a command fails, the container stops making any tracing impossible.
If you can't modify the container, then it's pretty difficult to analyse the issue.

This patch offer to "control" the verbosity of bash with a DEBUG environement variable.

If you need to trace a particular container, just set DEBUG to any value at run time like :
    sudo docker run -d --name ceph-mon -e DEBUG=1 -e NETWORK_AUTO_DETECT=1 daemon mon

In addition, this patch add an option to ask the container not to die is a command fails.

If 'DEBUG=stayalive' is set, the -x is enabled and the 'docker_exec.sh' file is sourced to
set trap on ERR and keep the container alive if a command fails (even the last exec call).