start_rgw: Avoid duplicated code
In this program, there is two occurences of duplicated code that over-complicate the reading.

This patch just take care of getting the conditional code being smaller
and then execute the command with arguments.

This patch is also adding an exec prior the final command as per docker rule.
The last process to be executed should be run in exec when using bash as init.