tools: Adding misslinks
When adding a file in the repository, it's pretty common to forget
adding the proper links to other versions :/

The misslinks tool helps at detecting that situation and report what
file is missing where.

You can pass as arguments the ceph_releases to check unless all of them are considered.

A typical output looks like :
	[user@host]./misslinks.sh  kraken/
	kraken//centos/7/daemon/osd_scenarios/osd_common.sh doesn't exists but found in kraken//ubuntu/16.04/daemon/osd_scenarios/osd_common.sh
	[user@host]

Note that misslinks doesn't consider if its a real file or a symlink.
It just check if a file with the same name exists across various OS & versions.