daemon: simply daemon startup command line
Prior to this change we were replicating the same command line to start
the ceph daemons. It is prone to errors while adding a new daemon so
instead we now use a variable that contains the command line arguments.
With this variable it is now easier to modify a command line, the change
will be propagated to all the daemons command line startup.

Signed-off-by: Sébastien Han <seb@redhat.com>