daemon: improve debug mode
This patch enhances the current support for debugging a container during
runtime.
This patch introduces the ability to specify more than one debug option.
We currently have 3 options when activating the DEBUG option:

  * DEBUG=verbose will enable 'set -x'
  * DEBUG=fstree along with an URL, so DEBUG=fstree=http://myfstree will
download the content of an URL (expecting an fs tree) and uncompress it
in /. This is useful if you want to test a patch and/or add new versions
of a binary.
  * DEBUG=stayalive will launch a shell when the container fails

You can combine these options using comma as separators like:

-e DEBUG=verbose,fstree=http://myfstree,stayalive

Signed-off-by: Sébastien Han <seb@redhat.com>