debug: Detecting root directory inside patched filesystem
When using a patching filesystem by using DEBUG=fs=<url> command line,
the root filesystem could be reallocated into a sub directory.

That's typically the case if you download a tarball from github.

This patch is trying to detect the root filesystem by searching the
first directory to feature a list of typical directories of a / like
/usr, /bin or similar.

Once it's found, the --strip= option of tar is used to skip the useless
directories.