debug: Implement failure proof loop
While looping on the tar -tf, it is expected that the egrep can fail.

In the current code, the egrep failed and exited immediately preventing the path detection code to work.

This patch is doing two things :
- avoid a global failure if the egrep fails
- Reverse counting to save the leftier directory that matched

Counting from left to right (seq 0 2) makes the code difficult to write
as it needs a break statement to avoid detecting the possible rightest
matches.

Reverting the seq (2 -1 0) only consider the leftier item without any break statement.