kubernetes: support jinja2 as alternative template processor
sigil is not available out-of-the-box (on Fedora that is), has a name
conflict with another project, and requires to setup a custom build tool
chain that didn't work for me. Use jinja2, as it should be available on
most systems and it's the template language used in, for example,
ansible and Django.

This might address https://github.com/ceph/ceph-docker/issues/338.