build_imgs.sh: simplify non-PR case in copy_dirs()
build_imgs.sh's copy_dirs() method is a no-op if the top-level "base",
"daemon", and "demo" directories already exist. We use this conditional
to determine what we're testing in Travis CI: are we testing a GitHub
PR, or a pushed branch?

Simplify the code here to simply exit the function early. This
eliminates a level of indentation so the function is easier to read and
understand.