daemon: do not run the monitor twice
First we remove the initial monmap, this monmap is really dangerous
after the initial deployment. Thus it should be deleted so we are sure
it's used later.

Also prior to this patch we were running the monitor which would fail
instantly, then create a socket and the other process would not run.
Removing this useless start and let's just the monitor start by reading
the Ceph configuration file, which will contain the list of all
monitors.

Also, we need to inject the map for kv or k8s env

This is needed for etcd or k8s deployments as new containers joining
need to have a map of the cluster.
The list of monitors will not be provided by the ceph.conf
since we don't have the overall knowledge of what's already deployed
In this kind of environment, the monmap is the only source
of truth for new monitor to attempt to join the existing quorum

Resolves: https://bugzilla.redhat.com/show_bug.cgi?id=1455357
Signed-off-by: Sébastien Han <seb@redhat.com>