daemon: fix dmcrypt list
We now use 'start_disk_list' to get the journal partition. The tool is
more reliable than ceph-disk list and the execution is faster.

Also we change the CLI_OPTS in the prepare sequence since it populates
options for ceph-disk, however this variable is later used by
start_disk_list to perform several actions. The subsequent command will
fail due to unknows options, e.g ceph --bluestore, this does not exist.
So now instead, just like the activation we build a specific command for ceph-disk prepare.

Signed-off-by: Sébastien Han <seb@redhat.com>