Convert `exit.sh` into `osd-exit.sh`, remove trap
Rename `exit.sh` to `osd-exit.sh`. The theory of operation has changed a
bit also. The script no longer tries to set a SIGTERM trap for use
generically in any orchestration environment, because the trap was being
deactivated after the entrypoint.sh script exited with success (or
failure). Instead, make this operate similarly to the `remove-mon.sh`
script. The script is still designed to gracefully unmount ceph mounts
in the container before terminating.

Signed-off-by: Blaine Gardner <blaine.gardner@suse.com>