daemon: remove OSD_FORCE_ZAP option
It is too dangerous to have this kind of capability, especially if there
is automation in place that uses this option.
Prior to deploy an OSD you must make sure the device is in a proper
shape. Also, ceph-disk prepare will create a GPT label if there is no
label on the drive.

Signed-off-by: Sébastien Han <seb@redhat.com>