osd: execute a post task after osd stops
We noticed that sometime, some devices are still mounted even after the
OSD has exited.
This commit ensures with a post action function all devices in the
container will be unmounted.

We define `sigterm_cleanup_post` inside the `osd_activate` itself
because we explicitely want to 'protect' the following `exec` and it's
nice to have the corresponding cleanup code at the same place.

Signed-off-by: Guillaume Abrioux <gabrioux@redhat.com>