docker_exec: ability to run pre/post tasks before exiting
The idea here is to be able to run pre and/or post actions before a
daemon exits.
This idea came with a recent case where we suspected than daemon could
let some leftover after exited (29302d1).

If needed to execute actions before and/or after the daemon exits, you
can implement function ideally in the entrypoint.

Here are the three functions you can implement :
- In case the daemon crashes  : `err_cleanup`

- In case the daemon receives SIGTERM signal and *BEFORE*
  the process is killed : `sigterm_cleanup_pre`

- In case the daemon receives SIGTERM signal and *AFTER*
  the process has been killed : `sigterm_cleanup_post`

To enter into these trap functions docker_exec.sh needs to become more generic.
Therefore, it is now loaded early in entrypoint.sh so we can use it.
This means we need to change the way the `stayalive` debug feature is called
by moving the call to `set_trap_err` in `debug.sh` intead.

Signed-off-by: Guillaume Abrioux <gabrioux@redhat.com>