disk_list.sh: improve 'ceph data' mount partition
We can assume the device is dmcrypt as soon as we find `TYPE=crypto_LUKS`
on `${OSD_DEVICE}`.
There is no need to try to predict what will be the name of the 'ceph data'
partition, therefore, `is_loop_dev` is no longer needed.
Let's find the `ceph data` uuid partition with `blkid` and mount it.

Fixes: #884

Signed-off-by: Guillaume Abrioux <gabrioux@redhat.com>