stage.py: Changing scanning order
To improve the visual output of the log file, it seems more accurate to process all for the directories for a given image at once.

The do_variable_replace() call is integrated in the main loop to insure that all the processing done on a particular image is completed before interating to
the next one.

With this patch, two adjascent lines have much more chance for being part of the same processed directory.