Add build staging integration test from POC #899
Based on the proof of concept from PR #899, add Python scripts to stage
files for building.

An integration test for this functionality can be found in
`tests/stage-test` and can be run as `make stage-test` or manually as
`./tests/stage-test/test_staging.sh`.

Additionally add flake8 configuration to tox.ini for linting the added
code. Currently, some of the existing code is also flagged by flake8, an
area for future improvement. flake8 linting can be run with `make lint`.

Signed-off-by: Blaine Gardner <blaine.gardner@suse.com>