common_function: make is_dmcrypt more generic
I adapted the function so I can use it with the zap "all_ceph_disks"
feature. The idea is to give the possibility to `is_dmcrypt` to take an
argument so we can loop over the whole list of devices in case of
zapping "all_ceph_disks", otherwise it default to the global
`${OSD_DEVICE}`.

Signed-off-by: Guillaume Abrioux <gabrioux@redhat.com>