kubernetes: add RBAC configuration
If RBAC is enabled on a cluster (which, for example, is the default on
clusters created with kubeadm) ceph pods need some extra permissions.

For now, we simply grant admin permissions to the default service
account in the ceph namespace. Since we assume that the ceph namespace
is exclusive to only ceph-related workloads this should be reasonably
safe.