kubernetes: rbd-provisioner deployment
Use the out-of-tree provisioner for the rbd storage class based on
https://github.com/kubernetes-incubator/external-storage/tree/master/ceph/rbd

Clusters created with kubeadm run the controller manager in the cluster
itself, but the image does not have the necessary tools to fulfill
kubernetes.io/rbd requests. Therefore, the out-of-tree provisioner is
required. See the included README for details.