Dockerfile: Adding tracability labels
When building containers, it's pretty easy to get many of them built in
a short time.

How can we identify if two users are using the same container and what could be the differences between both ?

This commit is about adding into the containers some labels that try to detail the build context with :
 - TIMESTAMP  : When does the build have been triggered ?
 - RELEASE    : What is the actual release ? If not defined, this equals the git branch name
 - GIT_REPO   : What was the url of the git repository
 - GIT_BRANCH : What was the git branch used to build this container
 - GIT_COMMIT : What was the commit ID of the current HEAD
 - GIT_CLEAN  : Was the repository clean when building ?

All that information should be useful to understand what was the context of the container build.