Split part of the readme into CONTRIBUTING.md
The readme after all of the great documentation we wrote for the new
project structure was huge. Split it into a readme useful for most
people who want to casually use/understand the project and a new
`CONTRIBUTING.md` aimed at developers.

Clean the CI section since it is no longer accurate with the thought
that this section should be filled in when the CI is finished being
reworked for the new project structure.

Signed-off-by: Blaine Gardner <blaine.gardner@suse.com>