staging: refactor envvars for required variables
Both DAEMON_BASE_IMAGE and DAEMON_BASE are required by staging, but they
are not checked in envvars to see if they're set.

While adding DAEMON_BASE_IMAGE and DAEMON_IMAGE, I realized that the end
goal of envvars is to ensure that required environment variables are
defined but that the  way it was used by staging made that
unclear. Adding new required variables was a confusing process. Refactor
staging's envglobals file to clarify its purpose and make adding
required variables in the future easier.

Because env vars are no longer imported by modules, refactoring this
also necessitated changing how logging is initialized and used.

Signed-off-by: Blaine Gardner <blaine.gardner@suse.com>