Makefile: Limiting build.parallel parallelism
When building multiple flavors in parallel with the build.parallel target, not specifying the number of jobs means unlimited.
Regarding the number of CPUs you have on your system, this could be very counter-productive.

This patch use the nproc commands to get the number of available CPUs, which could be limited by the system administrator of a build host.
This way we only build "nproc" jobs at a time to avoid a CPU overload.

Signed-off-by: Erwan Velu <evelu@redhat.com>