ci: Changing semantic of DO_NOT_PUSH_LATEST
DO_NOT_PUSH_LATEST is semanticly wrong here.
To be more generic, it would be better saying that we run on tagged head.
Then we let functions deciding what they have to do when we run on a tagged head.
Reading the code with this change shows a better semantic.

This commit replace DO_NOT_PUSH_LATEST by TAGGED_HEAD which is a boolean, false by default.
This commit also move the testing of TAGGED_HEAD inside push_ceph_imgs_latest() to get a cleaner main.
That also give the opportunity to print a specific & explicit message about this skip.

Signed-off-by: Erwan Velu <evelu@redhat.com>