Makefile: Ignore duplicate ids when cleaning
When cleaning the containers, the clean target search for docker images that match a particular regexep (daemon{-base})
It takes all of them and execute a rmi on it.

When tagging some of those images, like with the 'latest' tag used on
the CI, some images are reported twice in the list of images to delete.
The rmi tries to delete every image reported and fails on the duplicated one.

This patch simply make a unique on the output of image ids to avoid
such case. This way a duplicate idx have only a single occurence in our
list of images to delete.

Prior to this commit the following message was shown during the clean:

    Deleted: sha256:5d922b1dbe56c1a8b8076e79474e2dc8e44360f7a9146243ded768a82c76f24d
    Deleted: sha256:566483b5382d7217a6d24daf77b9e7533d5879a45e61f131cdf2d5c32dec310e
    Error response from daemon: No such image: cbdf2fdfed45:latest
    Error response from daemon: No such image: 0e34fefd5ad9:latest

Signed-off-by: Erwan Velu <evelu@redhat.com>