Makefile: Let user choose a specific version of Ceph
This patch is about adding an option to enforce a particular point release of ceph like "12.2.0" instead of the latest.

To enable this feature, the required ceph version have to be padded to the ceph release when specifying a flavor like in :

    make FLAVORS=luminous-12.2.2,centos,7 build

The separator is dash "-" for yum/dnf/zypper and equal "=" for apt.
The required version is saved into CEPH_POINT_RELEASE variable unless it remains empty.
If no version is given, package managers will take the latest one.

This patch only implements this feature for Centos as it will depends on distributions & repositories to offer/enable this feature.
Signed-off-by: Erwan Velu <evelu@redhat.com>