staging: Use an OrderedDict for `files_copied`
For an unknown reason, after the preceding changes, the order of the
output in `files-sources` is changed slightly. This could merely be the
result of the Python dict being unordered. Change files_copied to use an
`OrderedDict`, and change the ordering in the key value for
`files-sources` so that we know in the future that the ordering of these
files won't change arbitrarily.

Signed-off-by: Blaine Gardner <blaine.gardner@suse.com>