contrib: build latest-bis image
latest-bis is needed by ceph-ansible so it can test the restart
handlers on an image ID change.
So we rebuild latest again to get a different image ID.

We extracted the test for a tagged head so that we can separate each
function's call (push_ceph_imgs_latests and build_and_push_latest_bis).

latest-bis needs to exist if latest tags were built, so basically when
we are not building a release.

Signed-off-by: Sébastien Han <seb@redhat.com>