images: Use the tini entrypoint for both images
Install tini in daemon-base image, and set it as part of the entrypoint
in both images. tini should be a nice safeguard to protect
ceph-container from zombie processes (resource leak), and it ensures
signal handlers work properly.

tini source: https://github.com/krallin/tini

------------------------------------------------------------------------
Q&A

Q: Docker 1.13 has tini built in, so what gives adding it here?
A: In order for Docker to use tini, the `--init` flag must be passed,
   which Kubernetes does not do automatically, nor could I find an
   option to pass it. Outside of Kubernetes, adding it in the container
   directly also gets rid of user error forgetting to add the `--init`
   flag.

Q: In the daemon image, why specify a custom entrypoint with tini
   instead of running `entrypoint.sh` from a `CMD` instruction?
A: Specifying a custom entrypoint allows the user to issue a custom
   command and still have the container initialize as normal using the
   entrypoint. Otherwise, the user might have to execute
   `/entrypoint.sh` manually before going about what they set out to do
   in the first place. The entrypoint can be disabled/overridden
   manually if the user feels the need.