introduce post install tweaks
This commit introduces 'tweaks', tweak is a new capability that allows
you to configure the container OS. Typically this makes a particular
service more container aware or change its default configuration so it
can work better in containers.

Its first implementation is for configuring LVM.
lvm when creating devices tries to talk to udev. In a container this is
not possible and lvm can not reach udev on the host. Disabling the sync
and the rules help us getting through the lv creation.
Later to scan the lv on the host you can use lvs --readonly

Signed-off-by: Sébastien Han <seb@redhat.com>