mgr: check the admin key only if no mgr key
There is no need to fail if the admin key is not present especially
since other component might create the key (e.g: ansible).
We have to fail only the mgr key is not present and if the admin key
does not exist also as we will need it to create the mgr key.

Signed-off-by: Sébastien Han <seb@redhat.com>