make: rename REGISTRY to TAG_REGISTRY
There could be some confusion about whether `REGISTRY` refers to the
base registry from which the base image is pulled or whether it is the
repository with which the build images will be tagged. The intention is
that it is the latter. Clarify the confusion by changing `REGISTRY` to
`TAG_REGISTRY`.

Signed-off-by: Blaine Gardner <blaine.gardner@suse.com>