ceph-releases/rhel7: Adding compose script
The build mecanism for rhel is a bit specific and requires for now a
single docker file for the whole thing.

This generates an issue with the current project architecture as the
content is split across two stages (daemon/daemon-base).

This patch is about putting all the logic in daemon-base and copy the
remaining files & the end of the docker file from the daemon/ into a
composed directory.

The preparation of this directory is done via the
contrib/compose-rhcs.sh script. This is not really statisfying but will
work a workaround for a short.

Signed-off-by: Erwan Velu <evelu@redhat.com>