ceph-releases: Adding centos-arm64
This commit is about adding support for arm64 for centos.

In a perfect world, supporting a new arch is supposed to get anything
changed.

We are in a situation where the build of centos on ARM64 trigger the
following issues :
- nfs-ganesha & tcmu-runner packages are not available
- confd is not available from github

This commit is a little bit special as it try to workaround this
situation. It does create a centos-arm64 OS which links most of the
files from centos.

In this project, we are not supposed to have symlinks but as we are
using the same OS with a very few constrains specific to this arch,
Blaine & I considered that was the best approach.

The differences with centos are :

- __GANESHA_PACKAGES__ is empty to avoid it to be installed
- __WEB_INSTALL_CONFD__ is set to /bin/true to avoid the installation
- __ISCSI_PACKAGES__ is not defined to get it to empty to avoid the installation

The associated scenario are disabled : populate_kvstore kv_type_etcd kv_type_k8s nfs tcmu_runner rbd_target_api rbd_target_gw

This flavor have to be built in a particular fashion as centos-arm64 doesn't exist on the docker registry.
It have to be built with the following command :

    make FLAVORS=luminous,centos-arm64,7 BASE_IMAGE=centos:7 build

The BASE_IMAGE override the centos-arm64 to let docker use the centos:7 image.

Signed-off-by: Erwan Velu <evelu@redhat.com>