contrib: Adding build-push-ceph-container-imgs-arm64.sh
This script is a temporary hack to make arm64 builds to be sucessful.
It sources the original build-push-ceph-container-imgs.sh to share most
of the code with it and override a few specific functions.

The build process is particular because
- only centos/luminous is supported for now
- it needs to override the BASE_IMAGE during the make to fix the image name

This commit add a check in build-push-ceph-container-imgs.sh to check if
the following functions are already declared : build_ceph_imgs, push_ceph_imgs,
build_and_push_latest_bis. This allow the arm64 code to override the
functions of build-push-ceph-container-imgs.sh while keeping the other
functions and the same main().

Signed-off-by: Erwan Velu <evelu@redhat.com>