make: Refactor flavor BASEOS_ vars to DISTRO vars
It isn't very straightforward for a user that the BASEOS_REPO and
BASEOS_TAG specified in the flavor spec (`<ceph version>[point
release],<baseos repo>, <baseos tag>`) refer to the ceph-container
source directories. Fix this by changing BASEOS_REPO to DISTRO and
BASEOS_TAG to DISTRO_VERSION in the makefile. We still need BASEOS_REPO
and BASEOS_TAG, because we are planning to make these overrideable.

Signed-off-by: Blaine Gardner <blaine.gardner@suse.com>