make: Allow override of all BASEOS_ vars
Since we now are making the flavor clearly ceph-release source path
locations, the user may want to override any of the individual
components, not just BASE_IMAGE as a whole.

We still have BASE_IMAGE which can be used with `__ENV_[BASE_IMAGE]__`,
but BASE_IMAGE is now calculated and set in staging Python code. Since
visual feedback of the value is still valuable, we add a 'Computed:'
section to the staging output with BASE_IMAGE verified in it.

While updating the help text in the Makefile, I realized that the old
`@echo 'stuff'` syntax was cluttered, and it was easy to forget the
quote at the end of the line. Change to use a `define` block for the
text, which will allow us to specify anything without regard to whether
we have matched quotes, either single or double. Big win.

Signed-off-by: Blaine Gardner <blaine.gardner@suse.com>