ceph-releases/rhel7: temporarily hardcode release to "7"
In order to preserve upgrades from the GA container image, we need the
"release" label to be "7" (or higher).

Ordinarily OSBS will dynamically increment the "release" label to the
proper value (see
https://github.com/projectatomic/atomic-reactor/blob/master/atomic_reactor/plugins/pre_bump_release.py)

When the component name changes like this, OSBS will not know that the
last release was "6" in order to bump it to "7". We must hardcode
"release" for our first rhceph-rhel7-container build.

We can revert this commit when we have the first rhceph-rhel7-container
non-scratch build present in RH's downstream build system.